fix parsing of wide strings
This fixes https://github.com/n4r1b/ferrisetw/issues/60

There is no guarantee the wide string pointer is aligned, and thus
the vec of u16 values must be reconstructed, there is no way to avoid
it. The previous rejection on unaligned pointers led to failing to
parse valid values.